Lazy-require more modules to speed initial startup
Lazy-require hash-lang-bridge.rkt, package.rkt, and scribble.rkt.

On my machine, this causes racket--cmd-ready-p to go from reporting
over 8 seconds to under 5 seconds.

Of course, this only defers load time -- gets to "ready for any first
command" faster. When the back end is started /because of/ a command
needing one of those modules, that /total/ load time won't be
faster (and in fact could be slightly slower). Even so, I think this
is a good approach to get the back end up quickly, and not presume the
set of features actually utilized by any given user.